# api/openapi.yaml
openapi: 3.1.0
info:
  title: SlateDesk API
  description: |
    SlateDesk v2 HTTP API. Milestone 1 surface: local authentication
    (signed HttpOnly session cookie), users, and teams. Milestone 2 adds
    the ticket core: tickets, articles, attachments, tags, dashboard
    counters, and the realtime event stream. Milestone 3 adds the email
    channel's admin surface: mailbox configuration with live fetch/send
    tests and the Google OAuth connect flow, the instance external-URL
    setting, and outbound delivery status on articles with a retry
    operation.

    All errors are RFC 9457 problem details (`application/problem+json`).

    Access model (M2): every ticket-core operation requires an
    authenticated **agent or admin** session — customer access arrives
    with the M4 portal. The one stricter rule: `DELETE /tags/{id}` is
    **admin only**. Agents can additionally **read** users and teams
    (`GET /users`, `/users/{id}`, `/teams`, `/teams/{id}`) — the
    workspace needs them for assignee pickers, team routing, and
    requester lookups; all user/team mutations remain admin only. Each
    operation's description states its required role.

    Access model (M3): everything under `/mailboxes` and `/settings` is
    **admin only** (including the Google OAuth callback, which arrives as
    a top-level browser navigation carrying the admin's session cookie).
    `POST /articles/{id}/retry-send` is agent or admin, like the rest of
...

...
    (`/canned-replies`).

...

...

...
servers:
...
security:
  - cookieAuth: []
  - apiKey: []
...
paths:
  /auth/login:
    post:
      operationId: login
      summary: Log in with email and password
      description: |
        Verifies the credentials and, on success, sets the `sd_session`
        HttpOnly cookie (SameSite=Lax) carrying the signed session token.
        The response body is empty; the cookie is the session.
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "204":
          description: Authenticated. Session cookie set.
          headers:
            Set-Cookie:
              description: "`sd_session` signed HttpOnly session cookie."
              schema:
                type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          description: Invalid credentials or deactivated account.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "429":
          description: Too many login attempts; retry later.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        default:
          $ref: "#/components/responses/Problem"
  /auth/logout:
    post:
      operationId: logout
      summary: Log out
      description: |
        Clears the session cookie. Always succeeds for authenticated and
        unauthenticated callers alike; idempotent.
      tags: [auth]
      security: []
      responses:
        "204":
          description: Session cookie cleared.
          headers:
            Set-Cookie:
              description: Expired `sd_session` cookie.
              schema:
                type: string
        default:
          $ref: "#/components/responses/Problem"
  /auth/me:
    get:
      operationId: getCurrentUser
      summary: Get the authenticated user
      tags: [auth]
      responses:
        "200":
          description: The user the session cookie belongs to.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
        default:
          $ref: "#/components/responses/Problem"
  /users:
    get:
      operationId: listUsers
      summary: List users
      tags: [users]
      parameters:
        - name: role
          in: query
          description: Filter by role.
          schema:
            $ref: "#/components/schemas/Role"
        - name: active
          in: query
          description: Filter by active flag.
          schema:
            type: boolean
        - name: q
          in: query
          description: Case-insensitive substring match against email and name.
          schema:
            type: string
        - name: limit
          in: query
          description: Page size.
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
        - name: offset
          in: query
          description: Page offset.
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: Matching users, newest first.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        default:
          $ref: "#/components/responses/Problem"
    post:
      operationId: createUser
      summary: Create a user
      tags: [users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: A user with this email already exists.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        default:
          $ref: "#/components/responses/Problem"
  /users/{id}:
    parameters:
      - $ref: "#/components/parameters/UserID"
    get:
      operationId: getUser
      summary: Get a user
      tags: [users]
      responses:
        "200":
          description: The user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Problem"
    patch:
      operationId: updateUser
      summary: Update a user
      description: |
        Partial update; omitted fields are left unchanged. Setting `password`
        rehashes the password and bumps `token_version`, invalidating every
        outstanding session for the user. Email is immutable in M1.
      tags: [users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: The updated user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Problem"
    delete:
      operationId: deactivateUser
      summary: Deactivate a user
      description: |
        Soft delete: sets `active=false` and bumps `token_version` so all
        outstanding sessions die. The row is retained; there is no hard
        delete in M1.
      tags: [users]
      responses:
        "204":
          description: User deactivated.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Problem"
  /teams:
    get:
      operationId: listTeams
      summary: List teams
      tags: [teams]
      responses:
        "200":
          description: All teams, sorted by name.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Team"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        default:
          $ref: "#/components/responses/Problem"
    post:
      operationId: createTeam
      summary: Create a team
      tags: [teams]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTeamRequest"
      responses:
        "201":
          description: Team created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: A team with this name already exists.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        default:
          $ref: "#/components/responses/Problem"
  /teams/{id}:
    parameters:
      - $ref: "#/components/parameters/TeamID"
    get:
      operationId: getTeam
      summary: Get a team with its members
      tags: [teams]
      responses:
        "200":
          description: The team, including member users.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamWithMembers"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Problem"
    patch:
      operationId: updateTeam
      summary: Update a team
      description: Partial update; omitted fields are left unchanged.
      tags: [teams]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTeamRequest"
      responses:
        "200":
          description: The updated team.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: A team with this name already exists.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        default:
          $ref: "#/components/responses/Problem"
    delete:
      operationId: deleteTeam
      summary: Delete a team
      description: Hard delete. Membership rows cascade; users are untouched.
      tags: [teams]
      responses:
        "204":
          description: Team deleted.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Problem"
...
  /tickets:
    get:
      operationId: listTickets
      summary: List tickets
      description: |
        Filtered, paginated ticket list, most recent activity first
        (`updated_at` descending), or by relevance when `q` is present.
        Requires agent or admin role.

        `view` applies one of the fixed workspace views; every other
        filter composes with it (logical AND):

          - `my` — assigned to the caller, not closed
          - `unassigned` — no assignee, not closed
          - `open` — status is open, waiting_on_customer, or on_hold
          - `closed` — status is closed
      tags: [tickets]
      parameters:
        - name: view
          in: query
          description: Fixed workspace view applied before the other filters.
          schema:
            $ref: "#/components/schemas/TicketView"
        - name: status
          in: query
          description: Statuses to include (repeat the parameter). Empty = all.
          schema:
            type: array
            items:
              $ref: "#/components/schemas/TicketStatus"
        - name: priority
          in: query
          description: Priorities to include (repeat the parameter). Empty = all.
          schema:
            type: array
            items:
              $ref: "#/components/schemas/TicketPriority"
        - name: assignee_id
          in: query
          description: Only tickets assigned to this user.
          schema:
            type: string
            format: uuid
        - name: team_id
          in: query
          description: Only tickets owned by this team.
          schema:
            type: string
            format: uuid
        - name: tag_id
          in: query
          description: Only tickets bearing this tag.
          schema:
            type: string
            format: uuid
        - name: closed_today
          in: query
          description: |
            When true, only tickets closed since midnight (database
            timezone) — the same predicate as the dashboard's
            `closed_today` counter, so that tile can deep-link a queue
            whose contents match its number.
          schema:
            type: boolean
        - name: q
          in: query
          description: |
            Full-text search (websearch syntax, e.g. `printer -toner` or
            `"exact phrase"`) across ticket subjects and article bodies.
            When present, results are ranked by relevance.
          schema:
            type: string
        - name: limit
          in: query
          description: Page size.
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
        - name: offset
          in: query
          description: Page offset.
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: One page of matching tickets plus the total match count.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        default:
          $ref: "#/components/responses/Problem"
    post:
      operationId: createTicket
      summary: Create a ticket
      description: |
        Creates a ticket and its first article in one transaction: the
        day's ticket number (YYYYMMDD-NNNN) is allocated atomically, the
        body is recorded as an agent-authored public article
        (`sender_type=agent`, `channel=web`), and a `created` audit event
        plus a `ticket.created` stream event are emitted. The ticket
        starts `open`; priority defaults to `medium`. Requires agent or
        admin role.
      tags: [tickets]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTicketRequest"
      responses:
        "201":
          description: Ticket created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketDetail"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        default:
          $ref: "#/components/responses/Problem"
  /tickets/{id}:
    parameters:
      - $ref: "#/components/parameters/TicketID"
    get:
      operationId: getTicket
      summary: Get a ticket with its full thread
      description: |
        The ticket plus requester/assignee summaries, tags, every article
        in chronological order (each carrying its attachments), and the
        audit event trail. Requires agent or admin role.
      tags: [tickets]
      responses:
        "200":
          description: The ticket detail.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Problem"
    patch:
      operationId: updateTicket
      summary: Update ticket fields
      description: |
        Partial update of status, priority, assignee, and team. Omitted
        fields are left unchanged; an explicit `null` for `assignee_id`
        or `team_id` clears that field. Setting `status=open` on a closed
        ticket is the reopen path. Every change is audited and emits a
        `ticket.updated` stream event. Requires agent or admin role.
      tags: [tickets]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTicketRequest"
      responses:
        "200":
          description: The updated ticket.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Problem"
  /tickets/{id}/tags:
    put:
      operationId: setTicketTags
      summary: Replace a ticket's tags
      description: |
        Replaces the full tag set with `tag_ids` (idempotent PUT). Unknown
        tag ids fail the whole request with 400; nothing is applied
        partially. Emits a `tags_changed` audit event and a
        `ticket.updated` stream event. Requires agent or admin role.
      tags: [tickets]
      parameters:
        - $ref: "#/components/parameters/TicketID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetTicketTagsRequest"
      responses:
        "200":
          description: The ticket's new tag set, sorted by name.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tag"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Problem"
  /tickets/{id}/articles:
    post:
      operationId: createTicketArticle
      summary: Add an article to a ticket
      description: |
        Appends an agent-authored article (`sender_type=agent`,
        `channel=web`) — a public reply or, with `is_internal=true`, an
        internal note. `body_html`, when provided, is sanitized
        server-side (bluemonday) before storage. A public reply on an
        open ticket flips it to `waiting_on_customer`; internal notes
        never change status. Emits an `article_added` audit event and an
        `article.created` stream event. Requires agent or admin role.
      tags: [articles]
      parameters:
        - $ref: "#/components/parameters/TicketID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateArticleRequest"
      responses:
        "201":
          description: Article created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Problem"
  /articles/{id}/attachments:
    post:
      operationId: uploadArticleAttachment
      summary: Upload an attachment to an article
      description: |
        Multipart upload of one file (part name `file`), streamed to blob
        storage — the whole request body is capped at 25 MiB. The stored
        filename is the part's filename, sanitized; the content type is
        taken from the part header (falling back to
        `application/octet-stream`). Requires agent or admin role.
      tags: [attachments]
      parameters:
        - $ref: "#/components/parameters/ArticleID"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: The file to attach (max 25 MiB).
      responses:
        "201":
          description: Attachment stored.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attachment"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "413":
          $ref: "#/components/responses/ContentTooLarge"
        default:
          $ref: "#/components/responses/Problem"
  /articles/{id}/retry-send:
    post:
      operationId: retryArticleSend
      summary: Retry a failed outbound email delivery
      description: |
        Re-enqueues the email delivery of an outbound article whose
        `delivery_status` is `failed` (the terminal badge after the send
        job exhausted its retries). The article flips back to `queued`
        (emitting an `article.updated` stream event) and the send is
        retried with the **same** Message-ID that was committed before
        the first attempt, so a duplicate delivery can never fork the
        email thread. Only `failed` articles can be retried — anything
        else answers 409, with one recovery exception: an article stuck
        on `sending` whose send job no longer exists (the worker crashed
        on its final attempt) is also accepted; a send genuinely in
        flight still answers 409. Requires agent or admin role.
      tags: [articles]
      parameters:
        - $ref: "#/components/parameters/ArticleID"
      responses:
        "202":
          description: Delivery re-enqueued; the article is `queued` again.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: The article's delivery status is not `failed`.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        default:
          $ref: "#/components/responses/Problem"
...
  /dashboard/counters:
    get:
      operationId: getDashboardCounters
      summary: Get the dashboard counters
      description: |
        The four workspace counters in one call. `closed_today` counts
        tickets closed since midnight in the database's timezone (UTC in
        deployment). Requires agent or admin role.
      tags: [dashboard]
      responses:
        "200":
          description: Current counter values.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardCounters"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        default:
          $ref: "#/components/responses/Problem"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sd_session
      description: |
        Signed HttpOnly session cookie (SameSite=Lax) carrying the user id
        and token_version. Obtained via POST /auth/login; cleared via
        POST /auth/logout.
    apiKey:
      type: http
      scheme: bearer
      bearerFormat: sd_live_...
      description: |
        Scoped machine API key presented as
        `Authorization: Bearer sd_live_…`. An **alternative** to
        `cookieAuth` on every `/api/v1` operation: because the top-level
        `security` lists the two schemes as separate requirement objects
        (an OR), either scheme alone authenticates a request and the two
        are never both required. The key is a public `sd_live_` prefix
        plus 256 bits of base62 entropy; only its SHA-256 digest is
        stored, and the request acts as the admin or agent who created
        the key, carrying that key's scopes. Scopes are `read` and
        `write` — a `read` key is refused (403) on mutating operations.
        OpenAPI cannot bind scopes to an HTTP-bearer scheme, so the scope
        gate is enforced in the handler and documented rather than
        expressed as security scopes here.
...